Enum reustmann::instruction::Instruction

pub enum Instruction {
    Nop,
    Reset,
    Halt,
    In,
    Out,
    Pop,
    Dup,
    PushPc,
    PopPc,
    PopSp,
    SpTgt,
    PushNz,
    Swap,
    Push0,
    Add,
    Sub,
    Inc,
    Dec,
    Mul,
    Div,
    Xor,
    And,
    Or,
    Shl,
    Shr,
    Not,
    Bz,
    Bnz,
    Beq,
    Bgt,
    Blt,
    Bge,
    Loop,
    EndL,
    BraN,
    BraP,
    Target,
    Skip1,
    Skip2,
    Skip3,
    Skip4,
    Skip5,
    Skip6,
    Skip7,
    Skip8,
    Skip9,
}

These are the opcodes of the Reustmann instruction set, shown with their single-character mnemonics and long mnemonics.

Variants
Nop

No-operation, do-nothing

mnemonic: ;

The NOP opcode may be encoded in memory by the value zero, or by any value not assigned to another opcode. During execution, all unassigned opcode values are mapped to the NOP instruction.

PC = PC + 1 mod L
SP = no change
NZ = no change

Reset

Reset

mnemonic: R

PC = 0
SP = 0
NZ = false

Halt

Halt program execution

mnemonic: H

Causes program execution to stop.

In

Input a char from stdin, push it onto the stack

mnemonic: I

SP = SP - 1 mod L
*SP = getchar() trunc W
PC = PC + 1 mod L
NZ = true if the result stacked is nonzero, else false

Out

Pop a word from the stack, output to stdout

mnemonic: O

If the value on the top of the stack is outside the range of a char, it will be truncated to a char as it is output. This is inconsequential for Reustmann machines of rank W ≤ 8.

putchar((char)*SP)
SP = SP + 1 mod L
PC = PC + 1 mod L
NZ = true if the character output is nonzero, else false

Pop

Pop a word from the stack

mnemonic: p

SP = SP + 1 mod L
PC = PC + 1 mod L
NZ = true if the item popped is nonzero, else false

Dup

Duplicate the last stacked value

mnemonic: D

Temp = *SP
SP = SP - 1 mod L
*SP = Temp
PC = PC + 1 mod L
NZ = true if the value duplicated is nonzero, else false

PushPc

Push the PC onto the stack

mnemonic: C

SP = SP - 1 mod L
*SP = PC trunc W
PC = PC + 1 mod L
NZ = no change

PopPc

Pop the PC from the stack

mnemonic: c

PC = *SP mod L
SP = SP + 1 mod L
NZ = no change

PopSp

Pop the SP from the stack

mnemonic: Y

SP = *SP mod L
PC = PC + 1 mod L
NZ = no change

SpTgt

Set the SP to the next TARGET opcode

mnemonic: G

A search for the subsequent TARGET opcode is done at the time the SPTGT instruction is encountered, from the SPTGT instruction to memory location L - 1. The search does not wrap around. If no TARGET opcode is found, or if the PC is already at L - 1, the SPTGT is executed as if it were a NOP instruction.

If a subsequent TARGET opcode is found:
    SP = address of the TARGET opcode
else:
    SP = no change
PC = PC + 1 mod L
NZ = no change

PushNz

Push the NZ flag

mnemonic: P

SP = SP - 1 mod L
*SP = NZ
PC = PC + 1 mod L
NZ = no change

Swap

Swap the top two items on the stack

mnemonic: S

Temp = *SP
*SP = *(SP + 1 mod L)
*(SP + 1 mod L) = Temp
PC = PC + 1 mod L
NZ = no change

Push0

Push a zero onto the stack

mnemonic: 0

SP = SP - 1 mod L
*SP = 0
PC = PC + 1 mod L
NZ = false

Add

Add the top two stacked words, push the result

mnemonic: +

SP = SP - 1 mod L
*SP = (*(SP + 2 mod L) + *(SP+ 1 mod L)) trunc W
PC = PC + 1 mod L
NZ = true if the result is nonzero, else false

Sub

Subtract the top two stacked words and push the result

mnemonic: -

SP = SP - 1 mod L
*SP = (*(SP + 2 mod L) - *(SP + 1 mod L)) trunc W
PC = PC + 1 mod L
NZ = true if the result is nonzero, else false

Inc

Increment the item at the top of the stack

mnemonic: .

*SP = (*SP) + 1 trunc W
PC = PC + 1 mod L
NZ = true if the result is nonzero, else false

Dec

Decrement the item on the top of the stack

mnemonic: ,

*SP = (*SP) - 1 trunc W
PC = PC + 1 mod L
NZ = true if the result is nonzero, else false

Mul

Multiply the top two stacked words and push the result

mnemonic: *

SP = SP - 1 mod L
*SP= (*(SP + 2 mod L) * (*(SP + 1 mod L)) trunc W PC = PC + 1 mod L
NZ = true if the result is nonzero, else false

Div

Pop two words, divide, push the quotient and remainder

mnemonic: /

If the divisor is zero, the quotient will be the maximum possible word value, and the remainder zero.

Op0 = *(SP + 1 mod L)
Op1 = *SP
if Op1 is zero, change Op0 to the maximum value and Op1 to 1 *(SP + 1 mod L) = quotient of Op0 / Op1 trunc W
*SP = remainder of Op0 / Op1
PC = PC + 1 mod L
NZ = true if the quotient is nonzero, else false

Xor

Bitwise XOR the top two stacked words and push the result

mnemonic: ^

SP = SP - 1 mod L
*SP = (*(SP + 2 mod L) XOR *(SP + 2 mod L)) trunc W
PC = PC + 1 mod L
NZ = true if the result is nonzero, else false

And

Bitwise AND the top two stacked words and push the result

mnemonic: &

SP = SP - 1 mod L
*SP = (*( SP + 2 mod L) AND *(SP + 2 mod L)) trunc W
PC = PC + 1 mod L
NZ = true if the result is nonzero, else false

Or

Bitwise OR the top two stacked words and push the result

mnemonic: |

SP = SP - 1 mod L
*SP = (*(SP + 2 mod L) OR *(SP + 2 mod L)) trunc W
PC = PC + 1 mod L
NZ = true if the result is nonzero, else false

Shl

Logical shift left

mnemonic: (

*SP = *SP << 1 trunc W
PC = PC + 1 mod L
NZ = true if the result is nonzero, else false

Shr

Logical shift right

mnemonic: )

*SP = *SP >> 1 trunc W
PC = PC + 1 mod L
NZ = true if the result is nonzero, else false

Not

Bitwise NOT

mnemonic: ~

*SP = NOT *SP trunc W
PC = PC + 1 mod L
NZ = true if the result is nonzero, else false

Bz

Branch if zero (NZ flag is false)

mnemonic: Z

Skips one opcode if NZ is false.

if NZ is false
    PC = PC + 2 mod L
else
    PC = PC + 1 mod L
SP = no change NZ = no change

Bnz

Branch if nonzero (NZ flag is true)

mnemonic: z

Skips one opcode if NZ is true.

if NZ is true
    PC = PC + 2 mod L
else
    PC = PC + 1 mod L
SP = no change NZ = no change

Beq

Compare top two stacked words, branch if equal

mnemonic: =

if *(SP + 1 mod L) .eq. *SP
    PC = PC+ 2 mod L
else
    PC = PC + 1 mod L
SP = no change NZ = no change

Bgt

Compare top two stacked words, branch if greater than

mnemonic: >

if *(SP + 1 mod L) > *SP
    PC = PC + 2 mod L
else
    PC = PC + 1 mod L
SP = no change NZ = no change

Blt

Compare top two stacked words, branch if less than

mnemonic: {

if *(SP + 1 mod L) < *SP
    PC = PC + 2 mod L
else
    PC = PC + 1 mod L
SP = no change NZ = no change

Bge

Compare top two stacked words, branch if greater than or equal

mnemonic: }

if *(SP + 1 mod L) >= * SP
PC = PC + 2 mod L
else
    PC = PC + 1 mod L
SP = no change NZ = no change

Loop

Repeat the following instructions up to the next ENDL

mnemonic: L

PC = PC + 1 mod L
SP = no change
NZ = no change

EndL

End of LOOP

mnemonic: ]

Execution resumes at the instruction following the preceding LOOP opcode. A search for the preceding LOOP opcode is done at the time the ENDL instruction is encountered, from the current PC to location 0. The search does not wrap around. If no LOOP opcode is found, or if the PC is already at location 0, the ENDL is executed as if it were a NOP instruction.

if there is a preceding LOOP instruction:
    PC = location of LOOP opcode + 1
else:
    PC = PC + 1 mod L
SP = no change NZ = no change

BraN

Branch to the next TARGET opcode

mnemonic: B

A search for the subsequent TARGET opcode is done at the time the BRAN instruction is encountered, from the BRAN instruction to memory location L - 1. The search does not wrap around. If no TARGET opcode is found, the BRAN is executed as if it were a NOP instruction. If the TARGET is found at memory location L - 1, execution will resume at location 0.

if there is a subsequent TARGET instruction:
    PC = (location of TARGET opcode + 1) mod L
else:
    PC = PC + 1 mod L
SP = no change NZ = no change

BraP

Branch to the previous TARGET opcode

mnemonic: b

A search for the previous TARGET opcode is done at the time the BRAP instruction is encountered, from the BRAP instruction to memory location 0. The search does not wrap around.\ If no TARGET opcode is found or if the PC is already at location 0, the BRAP is executed as if it were a NOP. instruction.

if there is a prior TARGET instruction
    PC = location of TARGET opcode + 1
else
    PC = PC + 1 mod L
SP = no change NZ = no change

Target

Branch target for BRAN and BRAP

mnemonic: T

See SPTGT, BRAN, and BRAP instructions for the semantics. The TARGET opcode is just a marker, and is executed as if it were a NOP.

PC = PC + 1 mod L
SP = no change
NZ = no change

Skip1

Skip one instruction

mnemonic: 1

PC = PC + 2 mod L
SP = no change
NZ = no change

Skip2

Skip two instructions

mnemonic: 2

PC = PC + 3 mod L
SP = no change
NZ = no change

Skip3

Skip three instructions

mnemonic: 3

PC = PC + 4 mod L
SP = no change
NZ = no change

Skip4

Skip four instructions

mnemonic: 4

PC = PC + 5 mod L
SP = no change
NZ = no change

Skip5

Skip five instructions

mnemonic: 5

PC = PC + 6 mod L
SP = no change
NZ = no change

Skip6

Skip five instructions

mnemonic: 6

PC = PC + 7 mod L
SP = no change
NZ = no change

Skip7

Skip five instructions

mnemonic: 7

PC = PC + 8 mod L
SP = no change
NZ = no change

Skip8

Skip five instructions

mnemonic: 8

PC = PC + 9 mod L
SP = no change
NZ = no change

Skip9

Skip five instructions

mnemonic: 9

PC = PC + 10 mod L
SP = no change
NZ = no change